+++ Example/IGListKitExamples/ViewControllers/DemosViewController.swift
@@ -27,7 +27,8 @@ class DemosViewController: UIViewController, IGListAdapterDataSource {
         DemoItem(name: "Search Autocomplete", controllerClass: SearchViewController.self),
         DemoItem(name: "Mixed Data", controllerClass: MixedDataViewController.self),
         DemoItem(name: "Nested Adapter", controllerClass: NestedAdapterViewController.self),
-        DemoItem(name: "Empty View", controllerClass: EmptyViewController.self)
+        DemoItem(name: "Empty View", controllerClass: EmptyViewController.self),
+        DemoItem(name: "Single Section Controller", controllerClass: SingleSectionViewController.self)
     ]
 
     override func viewDidLoad() {