+++ Source/Internal/IGListAdapterInternal.h
@@ -18,8 +18,8 @@
 NS_ASSUME_NONNULL_BEGIN
 
 /// Generate a string representation of a reusable view class when registering with a UICollectionView.
-NS_INLINE NSString *IGListReusableViewIdentifier(Class viewClass, NSString * _Nullable kind) {
-  return [NSString stringWithFormat:@"%@%@", kind ?: @"", NSStringFromClass(viewClass)];
+NS_INLINE NSString *IGListReusableViewIdentifier(Class viewClass, NSString * _Nullable nibName, NSString * _Nullable kind) {
+    return [NSString stringWithFormat:@"%@%@%@", kind ?: @"", nibName ?: @"", NSStringFromClass(viewClass)];
 }
 
 @interface IGListAdapter ()
@@ -54,6 +54,7 @@ IGListCollectionContext
 @property (nonatomic, strong, nullable) IGListSectionMap *previoussectionMap;
 
 @property (nonatomic, strong) NSMutableSet<Class> *registeredCellClasses;
+@property (nonatomic, strong) NSMutableSet<NSString *> *registeredNibNames;
 @property (nonatomic, strong) NSMutableSet<NSString *> *registeredSupplementaryViewIdentifiers;
 
 - (NSArray *)indexPathsFromSectionController:(IGListSectionController <IGListSectionType> *)sectionController